feat: add search field in category picker
