+++ lib/features/calendar/ui/widgets/time_by_category_chart_legend.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:lotti/features/calendar/state/day_view_controller.dart';
+import 'package:lotti/features/categories/ui/widgets/categories_type_card.dart';
 import 'package:lotti/l10n/app_localizations_context.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/date_utils_extension.dart';
 import 'package:lotti/widgets/charts/utils.dart';
-import 'package:lotti/widgets/settings/categories/categories_type_card.dart';
 
 class TimeByCategoryChartLegend extends ConsumerWidget {
   const TimeByCategoryChartLegend({
