+++ lib/features/categories/ui/widgets/select_color_field.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_colorpicker/flutter_colorpicker.dart';
+import 'package:lotti/features/categories/ui/widgets/categories_type_card.dart';
 import 'package:lotti/l10n/app_localizations_context.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/color.dart';
-import 'package:lotti/widgets/settings/categories/categories_type_card.dart';
 
 class SelectColorField extends StatefulWidget {
   const SelectColorField({
