+++ lib/features/dashboards/ui/widgets/dashboards_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:lotti/classes/entity_definitions.dart';
-import 'package:lotti/widgets/settings/categories/categories_type_card.dart';
+import 'package:lotti/features/categories/ui/widgets/categories_type_card.dart';
 import 'package:lotti/widgets/settings/settings_card.dart';
 
 class DashboardCard extends StatelessWidget {
