+++ lib/features/tasks/ui/time_recording_icon.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:lotti/classes/journal_entities.dart';
+import 'package:lotti/features/categories/ui/widgets/categories_type_card.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/time_service.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/widgets/settings/categories/categories_type_card.dart';
 
 class TimeRecordingIcon extends StatelessWidget {
   const TimeRecordingIcon({
