+++ lib/l10n/app_en.arb
@@ -29,6 +29,7 @@
   "cancelButton": "Cancel",
   "categoryDeleteConfirm": "YES, DELETE THIS CATEGORY",
   "categoryDeleteQuestion": "Do you want to delete this category?",
+  "categorySearchPlaceholder": "Search categories...",
   "checklistAddItem": "Add a new item",
   "checklistItemDelete": "Delete checklist item?",
   "checklistDelete": "Delete checklist?",
